added doxygen from cmake (#239)
* added doxygen from cmake

* renaming/organize

* added 2d doxygen comment

* doxygen for concurrent 2d

* doxygen for the scheduler

* doxygen for tetmesh

* doxygen for ConcurrentTetMesh

* remove the comments for the new implementation

* removed some duplicate functions

* mainpage setup

* Apply suggestions from code review

* Create doxygen.yml

* Update doxygen.yml

* try doxgen

* make
