+++ CMakeLists.txt
@@ -1,5 +1,6 @@
 # Detects whether this is a top-level project
 get_directory_property(HAS_PARENT PARENT_DIRECTORY)
+
 if(HAS_PARENT)
     set(WILDMESHING_TOOLKIT_TOPLEVEL_PROJECT OFF)
 else()
@@ -8,6 +9,7 @@ endif()
 
 # Check required CMake version
 set(REQUIRED_CMAKE_VERSION "3.15.0")
+
 if(WILDMESHING_TOOLKIT_TOPLEVEL_PROJECT)
     cmake_minimum_required(VERSION ${REQUIRED_CMAKE_VERSION})
 else()
@@ -17,13 +19,14 @@ else()
     endif()
 endif()
 
-################################################################################
+# ###############################################################################
 project(WildMeshingToolkit DESCRIPTION "A mesh optimization toolkit")
-################################################################################
+
+# ###############################################################################
+option(WMTK_BUILD_DOCS "Build doxygen" OFF)
 
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake/recipes/")
-
 
 include(sanitizers)
 include(wmtk_colors)
@@ -37,9 +40,9 @@ set_property(GLOBAL PROPERTY USE_FOLDERS ON)
 # Generate position independent code by default
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 
-################################################################################
+# ###############################################################################
 # WildMeshingToolkit library
-################################################################################
+# ###############################################################################
 
 # Core library
 file(GLOB_RECURSE SRC_FILES CONFIGURE_DEPENDS
@@ -66,12 +69,14 @@ target_compile_features(wildmeshing_toolkit PUBLIC cxx_std_17)
 source_group(TREE "${CMAKE_CURRENT_SOURCE_DIR}/src" PREFIX "Source Files" FILES ${SRC_FILES})
 
 # Dependencies
-if (NOT TARGET igl::core)
+if(NOT TARGET igl::core)
     include(libigl)
 endif()
-if (NOT TARGET spdlog)
+
+if(NOT TARGET spdlog)
     include(spdlog)
 endif()
+
 include(geogram)
 include(onetbb)
 include(mshio)
@@ -90,9 +95,9 @@ target_link_libraries(wildmeshing_toolkit PUBLIC
     Tracy::TracyClient
 )
 
-################################################################################
+# ###############################################################################
 # Subdirectories
-################################################################################
+# ###############################################################################
 
 # Compile extras only if this is a top-level project
 if(WILDMESHING_TOOLKIT_TOPLEVEL_PROJECT)
@@ -102,4 +107,28 @@ if(WILDMESHING_TOOLKIT_TOPLEVEL_PROJECT)
 
     # Demo apps
     add_subdirectory(app)
+endif()
+
+if(WMTK_BUILD_DOCS)
+    # check if Doxygen is installed
+    find_package(Doxygen)
+
+    if(DOXYGEN_FOUND)
+        # set input and output files
+        set(DOXYGEN_IN ${CMAKE_CURRENT_SOURCE_DIR}/docs/Doxyfile.in)
+        set(DOXYGEN_OUT ${CMAKE_CURRENT_BINARY_DIR}/Doxyfile)
+
+        # request to configure the file
+        configure_file(${DOXYGEN_IN} ${DOXYGEN_OUT} @ONLY)
+        message("Doxygen build started")
+
+        # note the option ALL which allows to build the docs together with the application
+        add_custom_target(wmtk_doc_doxygen ALL
+            COMMAND ${DOXYGEN_EXECUTABLE} ${DOXYGEN_OUT}
+            WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
+            COMMENT "Generating API documentation with Doxygen"
+            VERBATIM)
+    else(DOXYGEN_FOUND)
+        message("Doxygen need to be installed to generate the doxygen documentation")
+    endif(DOXYGEN_FOUND)
 endif()
