+++ app/qslim/tests/CMakeLists.txt
