+++ app/shortest_edge_collapse/CMakeLists.txt
