+++ app/shortest_edge_collapse/app/CMakeLists.txt
