+++ app/shortest_edge_collapse/src/sec/envelope/SampleEnvelope.hpp
