+++ app/shortest_edge_collapse/tests/CMakeLists.txt
