+++ app/shortest_edge_collapse/tests/main.cpp
